Remove the static context and items not related to the library, and ease the restrictions on the bounds for an Atlas.
I am aware a lot of interfaces were removed, and it was because I made the decision to omit such models and a static context from the library. I still want data validations, but some needed to be altered, so I don't have any magical numbers floating around. Having a singleton was too problematic for testing, and broke some tests around the project when the tests were updated to account for a null configuration.