Remove 2 classes and migrate their methods to an "AssertionHelper"
The logic-checking helper methods from "StringHelper", "ObjectHelper" and "MathHelper" were moved to a single class, which removed them. I added more tests to cover most (hopefully all) of the logic, and also add a ReflectionHelper to assist with debugging. The logging examples have been restructured to have a more uniform method. I'm going to find the rest of the areas where an exception is thrown, though.