Adjust signature construction and update any instances of a "descriptor" or "message" being passed to the AssertionHelper methods
This entails quite a few changes, but it simplifies the need to construct a signature or message, provided the AssertionHelper has a method that aligns with the calling method's needs.

Also, the AssertionHelper has been called in more places, and throughout the project, explicit types have been used instead of 'var', when it isn't fully clear what type is used.