Add another test and a new implementation
The one before didn't work, and I opted to write a more performant one as compared to before. Also, the logging included is just being saved as a future reference. I'm still pondering if I want to supply logging for testing, or if it'll just be as I need it.