Add current implementation to the HexagonalNetworkGenerator
Note: It doesn't fully work. It only works for one case, and I have to either tinker with it more, or redo it again. More than likely, I'll simplify it, and do it in a better way, since this way is overly cumbersome. It also doesn't handle them alternating rows, but I have another idea of how it'll work.